+++ R/geary.R
@@ -101,7 +101,7 @@ geary.mc <- function(x, listw, nsim, zero.policy=NULL,
             zero.policy <- get("zeroPolicy", envir = .spdepOptions)
         stopifnot(is.logical(zero.policy))
         stopifnot(is.vector(x))
-	alternative <- match.arg(alternative, c("less", "greater"))
+	alternative <- match.arg(alternative, c("less", "greater", "two.sided"))
 	if(!inherits(listw, "listw")) stop(paste(deparse(substitute(listw)),
 		"is not a listw object"))
 	if(!is.numeric(x)) stop(paste(deparse(substitute(x)),
@@ -155,6 +155,8 @@ geary.mc <- function(x, listw, nsim, zero.policy=NULL,
         	pval <- punif((diff + 1)/(nsim + 1), lower.tail=FALSE)
     	else if (alternative == "less") 
         	pval <- punif((diff + 1)/(nsim + 1))
+        else pval <- punif(abs(xrank - (nsim+1)/2)/(nsim + 1), 0, 0.5,
+                lower.tail=FALSE)
 	if (!is.finite(pval) || pval < 0 || pval > 1) 
 		warning("Out-of-range p-value: reconsider test arguments")
 	statistic <- res[nsim+1]