+++ R/jc.R
@@ -109,7 +109,7 @@ print.jclist <- function(x, ...) {
 
 joincount.mc <- function(fx, listw, nsim, zero.policy=FALSE,
 	alternative="greater", spChk=NULL) {
-	alternative <- match.arg(alternative, c("greater", "less"))
+	alternative <- match.arg(alternative, c("greater", "less", "two.sided"))
 	if(!inherits(listw, "listw")) stop(paste(deparse(substitute(listw)),
 		"is not a listw object"))
 	if(!is.factor(fx)) stop(paste(deparse(substitute(fx)),
@@ -156,6 +156,8 @@ joincount.mc <- function(fx, listw, nsim, zero.policy=FALSE,
         		pval <- punif((diff + 1)/(nsim + 1), lower.tail=FALSE)
     		else if (alternative == "greater") 
         		pval <- punif((diff + 1)/(nsim + 1))
+                else pval <- punif(abs(xrank[i] - (nsim+1)/2)/(nsim + 1), 
+                    0, 0.5, lower.tail=FALSE)
 		if (!is.finite(pval) || pval < 0 || pval > 1) 
 		    warning("Out-of-range p-value: reconsider test arguments")