+++ NEWS.md
@@ -1,6 +1,8 @@
 # Version 1.2-6 (development)
 
-* #92, #93, #94 contributions of prototype bivariate Moran, local  bivariate Moran and local joincount tests by Josiah Parry
+* #95 add `"two.sided"` to `lee.mc()`, same for `sp.mantel.mc()`, `EBImoran.mc()`, `joincount.mc()`, `geary.mc()`
+
+* #92, #93, #94, #96  contributions of prototype bivariate Moran, local  bivariate Moran and local joincount tests by Josiah Parry
 
 * #91 `tolerance.nb()` update by F. Guillaume Blanchet